refactor: User 엔티티를 auth/entity에서 global/entity로 이동
