+++ src/main/java/com/upstage/devup/user/account/dto/UserAccountDto.java
@@ -1,6 +1,6 @@
 package com.upstage.devup.user.account.dto;
 
-import com.upstage.devup.auth.domain.entity.User;
+import com.upstage.devup.global.entity.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
