+++ src/main/java/com/upstage/devup/user/answer/repository/AnswerUserRepository.java
@@ -1,6 +1,6 @@
 package com.upstage.devup.user.answer.repository;
 
-import com.upstage.devup.auth.domain.entity.User;
+import com.upstage.devup.global.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AnswerUserRepository extends JpaRepository<User, Long> {
